WorkshopDemo/pyproject.toml
[project]
name = "hybrid-rag-agent"
version = "0.1.0"
description = "Hybrid RAG Knowledge Graph Agent with Pydantic AI"
requires-python = ">=3.11"
dependencies = [
    "pydantic-ai>=0.0.49",
    "pydantic-settings>=2.0.0",
    "asyncpg>=0.29.0",
    "graphiti-core>=0.5.0",
    "neo4j>=5.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
